refactor: extract shared helpers into _utils.py
Moves _LINESTYLE_ALIASES, _arr_to_b64, _norm_linestyle, _normalize_image,
_CMAP_ALIASES, _build_colormap_lut, and _resample_mesh out of figure_plots.py
into a new anyplotlib/_utils.py. Backward-compat re-imports are added to
figure_plots.py so all existing callers continue to work unchanged.
